move tax class to invoice namespace
