+++ tests/Entity/InvoiceTemplateTest.php
@@ -13,8 +13,8 @@ use App\Entity\Customer;
 use App\Entity\CustomerMeta;
 use App\Entity\InvoiceTemplate;
 use App\Entity\InvoiceTemplateMeta;
-use App\Entity\Tax;
 use App\Entity\TaxType;
+use App\Invoice\Tax;
 use Doctrine\Common\Collections\Collection;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
